+++ misc/Scrapix.postman_collection.json
@@ -92,7 +92,7 @@
 				"header": [],
 				"body": {
 					"mode": "raw",
-					"raw": "{\n    \"urls\": [\n        \"https://meilisearch.com/docs\",\n        \"https://www.meilisearch.com/docs\"\n    ],\n    \"meilisearch_url\": \"{{meilisearch_url}}\",\n    \"meilisearch_api_key\": \"{{meilisearch_api_key}}\",\n    \"meilisearch_index_uid\": \"{{meilisearch_index_uid}}\",\n    \"strategy\": \"docsearch\"\n}",
+					"raw": "{\n    \"urls\": [\n        \"https://meilisearch.com/docs\",\n        \"https://www.meilisearch.com/docs\"\n    ],\n    \"meilisearch_url\": \"{{meilisearch_url}}\",\n    \"meilisearch_api_key\": \"{{meilisearch_api_key}}\",\n    \"meilisearch_index_uid\": \"{{meilisearch_index_uid}}\",\n    \"strategy\": \"docssearch\"\n}",
 					"options": {
 						"raw": {
 							"language": "json"
